move the 'create_module' fixture into 'test_registry'
